wordclock: Replace the old usermod
