+++ usermods/usermod_v2_word_clock/usermod_v2_word_clock.cpp
@@ -1,507 +1,429 @@
 #include "wled.h"
 
+#if defined(WORD_CLOCK_LANGUAGE_NL)
+  #include "lang/word_clock_language_nl.h"
+  namespace WordClock = WordClockDutch;
+#else
+  #include "lang/word_clock_language_de.h"
+  namespace WordClock = WordClockGerman;
+#endif
+
 /*
- * Usermods allow you to add own functionality to WLED more easily
- * See: https://github.com/wled-dev/WLED/wiki/Add-own-functionality
+ * Word Clock Usermod
+ * This usermod displays the time in words using a compile-time selected
+ * language pack. It uses the WLED V2 usermod API to integrate with WLED and
+ * apply an overlay on the LED strip to light up the characters representing
+ * the current time in Dutch. This assumes that each LED of the LED strip is
+ * arranged behind a matrix of characters, lighting up one character per LED,
+ * and that the character matrix contains all words needed to display the time
+ * in a language-specific sentence.
  * 
- * This usermod can be used to drive a wordclock with a 11x10 pixel matrix with WLED. There are also 4 additional dots for the minutes. 
- * The visualisation is described in 4 mask with LED numbers (single dots for minutes, minutes, hours and "clock/Uhr").
- * There are 2 parameters to change the behaviour:
- * 
- * active: enable/disable usermod
- * diplayItIs: enable/disable display of "Es ist" on the clock.
+ * These settings are available in the Config > Usermods > Word Clock
+ * Settings page:
+ *
+ *   - `Active`: turn the word clock on or off.
+ *   - `Brightness Active`: brightness of the letters used for the current
+ *     time. Use 0 for off and 255 for full brightness.
+ *   - `Brightness Inactive`: brightness of the other letters. Use 0 for off
+ *     and 255 for full brightness.
+ *   - `Meander`: set to `false` when the LED strip runs left to right on
+ *      every row. Set to `true` when each row alternates direction.
+ *   - `Character Matrix`: the uppercase letters in your clock face. Include
+ *      all the words needed to display the Dutch time sentences, with each
+ *      row placed directly after the previous row.
+ *   - `Character Matrix Width`: the number of letters in each row. It cannot
+ *     be greater than the total number of letters or smaller than the longest
+ *     word the clock needs to display.
+ *   - `Matrix Char Offset`: the number of letters to skip at the beginning of
+ *     the matrix when matching letters to LEDs. This is useful when the first
+ *     physical LEDs do not correspond to the first letters.
+ *   - `Test Hour`: the hour to display for testing, from 0 to 23. Set it to
+ *     -1 to use the real time.
+ *   - `Test Minute`: the minute to display for testing, from 0 to 59.
  */
 
-class WordClockUsermod : public Usermod 
+class WordClockUsermod : public Usermod
 {
-  private:
-    unsigned long lastTime = 0;
-    int lastTimeMinutes = -1;
+private:
+  // The matrix of characters that can be highlighted to display the time.
+  // This matrix must contain all words needed to display the time in a
+  // Dutch sentence, e.g. "HET IS KWART OVER TIEN".
+  // The characters of each row are stored sequentially, and the rows are
+  // stored sequentially as well, left to right, and top to bottom.
+  String characterMatrix = FPSTR(WordClock::DEFAULT_CHARACTER_MATRIX);
 
-    // set your config variables to their boot default value (this can also be done in readFromConfig() or a constructor if you prefer)
-    bool usermodActive = false;
-    bool displayItIs = false;
-    int ledOffset = 100;
-    bool meander = false;
-    bool nord = false;
-    
-    // defines for mask sizes
-    #define maskSizeLeds        114
-    #define maskSizeMinutes     12
-    #define maskSizeMinutesMea  12
-    #define maskSizeHours       6
-    #define maskSizeHoursMea    6
-    #define maskSizeItIs        5
-    #define maskSizeMinuteDots  4
+  // The number of characters per row
+  int characterMatrixWidth = WordClock::DEFAULT_CHARACTER_MATRIX_WIDTH;
 
-    // "minute" masks
-    // Normal wiring
-    const int maskMinutes[14][maskSizeMinutes] = 
-    {
-      {107, 108, 109,  -1,  -1,  -1,  -1,  -1,  -1,  -1,  -1,  -1}, // 0 - 00
-      {  7,   8,   9,  10,  40,  41,  42,  43,  -1,  -1,  -1,  -1}, // 1 - 05 fünf nach
-      { 11,  12,  13,  14,  40,  41,  42,  43,  -1,  -1,  -1,  -1}, // 2 - 10 zehn nach
-      { 26,  27,  28,  29,  30,  31,  32,  -1,  -1,  -1,  -1,  -1}, // 3 - 15 viertel
-      { 15,  16,  17,  18,  19,  20,  21,  40,  41,  42,  43,  -1}, // 4 - 20 zwanzig nach
-      {  7,   8,   9,  10,  33,  34,  35,  44,  45,  46,  47,  -1}, // 5 - 25 fünf vor halb
-      { 44,  45,  46,  47,  -1,  -1,  -1,  -1,  -1,  -1,  -1,  -1}, // 6 - 30 halb
-      {  7,   8,   9,  10,  40,  41,  42,  43,  44,  45,  46,  47}, // 7 - 35 fünf nach halb
-      { 15,  16,  17,  18,  19,  20,  21,  33,  34,  35,  -1,  -1}, // 8 - 40 zwanzig vor
-      { 22,  23,  24,  25,  26,  27,  28,  29,  30,  31,  32,  -1}, // 9 - 45 dreiviertel
-      { 11,  12,  13,  14,  33,  34,  35,  -1,  -1,  -1,  -1,  -1}, // 10 - 50 zehn vor
-      {  7,   8,   9,  10,  33,  34,  35,  -1,  -1,  -1,  -1,  -1}, // 11 - 55 fünf vor
-      { 26,  27,  28,  29,  30,  31,  32,  40,  41,  42,  43,  -1}, // 12 - 15 alternative viertel nach
-      { 26,  27,  28,  29,  30,  31,  32,  33,  34,  35,  -1,  -1}  // 13 - 45 alternative viertel vor
-    };
+  // Is the ledstrip always from left to right on each row, or does it
+  // meander through the rows (i.e. go from left to right on the first row,
+  // then continue right to left on the second row, and so on)?
+  bool meander = true;
 
-    // Meander wiring
-    const int maskMinutesMea[14][maskSizeMinutesMea] = 
-    {
-      { 99, 100, 101,  -1,  -1,  -1,  -1,  -1,  -1,  -1,  -1,  -1}, // 0 - 00
-      {  7,   8,   9,  10,  33,  34,  35,  36,  -1,  -1,  -1,  -1}, // 1 - 05 fünf nach
-      { 18,  19,  20,  21,  33,  34,  35,  36,  -1,  -1,  -1,  -1}, // 2 - 10 zehn nach
-      { 26,  27,  28,  29,  30,  31,  32,  -1,  -1,  -1,  -1,  -1}, // 3 - 15 viertel
-      { 11,  12,  13,  14,  15,  16,  17,  33,  34,  35,  36,  -1}, // 4 - 20 zwanzig nach
-      {  7,   8,   9,  10,  41,  42,  43,  44,  45,  46,  47,  -1}, // 5 - 25 fünf vor halb
-      { 44,  45,  46,  47,  -1,  -1,  -1,  -1,  -1,  -1,  -1,  -1}, // 6 - 30 halb
-      {  7,   8,   9,  10,  33,  34,  35,  36,  44,  45,  46,  47}, // 7 - 35 fünf nach halb
-      { 11,  12,  13,  14,  15,  16,  17,  41,  42,  43,  -1,  -1}, // 8 - 40 zwanzig vor
-      { 22,  23,  24,  25,  26,  27,  28,  29,  30,  31,  32,  -1}, // 9 - 45 dreiviertel
-      { 18,  19,  20,  21,  41,  42,  43,  -1,  -1,  -1,  -1,  -1}, // 10 - 50 zehn vor
-      {  7,   8,   9,  10,  41,  42,  43,  -1,  -1,  -1,  -1,  -1}, // 11 - 55 fünf vor
-      { 26,  27,  28,  29,  30,  31,  32,  33,  34,  35,  36,  -1}, // 12 - 15 alternative viertel nach
-      { 26,  27,  28,  29,  30,  31,  32,  41,  42,  43,  -1,  -1}  // 13 - 45 alternative viertel vor
-    };
+  // Keep track of the last time our loop executed.
+  // Initialised to trigger an update on the very first loop() call.
+  unsigned long lastTime = ULONG_MAX - 60000UL;
+
+  // ledMask[i] is true if the LED at index i should be on for the current time, and false if it should be off
+  bool* ledMask = nullptr;
 
 
-    // hour masks
-    // Normal wiring
-    const int maskHours[13][maskSizeHours] = 
-    {
-      { 55,  56,  57,  -1,  -1,  -1}, // 01: ein
-      { 55,  56,  57,  58,  -1,  -1}, // 01: eins
-      { 62,  63,  64,  65,  -1,  -1}, // 02: zwei
-      { 66,  67,  68,  69,  -1,  -1}, // 03: drei
-      { 73,  74,  75,  76,  -1,  -1}, // 04: vier
-      { 51,  52,  53,  54,  -1,  -1}, // 05: fünf
-      { 77,  78,  79,  80,  81,  -1}, // 06: sechs
-      { 88,  89,  90,  91,  92,  93}, // 07: sieben
-      { 84,  85,  86,  87,  -1,  -1}, // 08: acht
-      {102, 103, 104, 105,  -1,  -1}, // 09: neun
-      { 99, 100, 101, 102,  -1,  -1}, // 10: zehn
-      { 49,  50,  51,  -1,  -1,  -1}, // 11: elf
-      { 94,  95,  96,  97,  98,  -1}  // 12: zwölf and 00: null
-    };
-    // Meander wiring
-    const int maskHoursMea[13][maskSizeHoursMea] = 
-    {
-      { 63,  64,  65,  -1,  -1,  -1}, // 01: ein
-      { 62,  63,  64,  65,  -1,  -1}, // 01: eins
-      { 55,  56,  57,  58,  -1,  -1}, // 02: zwei
-      { 66,  67,  68,  69,  -1,  -1}, // 03: drei
-      { 73,  74,  75,  76,  -1,  -1}, // 04: vier
-      { 51,  52,  53,  54,  -1,  -1}, // 05: fünf
-      { 83,  84,  85,  86,  87,  -1}, // 06: sechs
-      { 88,  89,  90,  91,  92,  93}, // 07: sieben
-      { 77,  78,  79,  80,  -1,  -1}, // 08: acht
-      {103, 104, 105, 106,  -1,  -1}, // 09: neun
-      {106, 107, 108, 109,  -1,  -1}, // 10: zehn
-      { 49,  50,  51,  -1,  -1,  -1}, // 11: elf
-      { 94,  95,  96,  97,  98,  -1}  // 12: zwölf and 00: null
-    };
+  // Set your config variables to their boot default value (this can also be done in readFromConfig() or a constructor if you prefer)
 
-    // mask "it is"
-    const int maskItIs[maskSizeItIs] = {0, 1, 3, 4, 5};
+  // Is this usermod active?
+  bool usermodActive = false;
 
-    // mask minute dots
-    const int maskMinuteDots[maskSizeMinuteDots] = {110, 111, 112, 113};
+  bool displayItIs = false;
+  bool nord = false;
+  int ledOffset =
+#if defined(WORD_CLOCK_LANGUAGE_NL)
+    0;
+#else
+    100;
+#endif
 
-    // overall mask to define which LEDs are on
-    int maskLedsOn[maskSizeLeds] = 
-    {
-      0,0,0,0,0,0,0,0,0,0,0,
-      0,0,0,0,0,0,0,0,0,0,0,
-      0,0,0,0,0,0,0,0,0,0,0,
-      0,0,0,0,0,0,0,0,0,0,0,
-      0,0,0,0,0,0,0,0,0,0,0,
-      0,0,0,0,0,0,0,0,0,0,0,
-      0,0,0,0,0,0,0,0,0,0,0,
-      0,0,0,0,0,0,0,0,0,0,0,
-      0,0,0,0,0,0,0,0,0,0,0,
-      0,0,0,0,0,0,0,0,0,0,0,
-      0,0,0,0
-    };
+  // Opacity (0=off, 255=full brightness) applied to LEDs that ARE part of the current time sentence.
+  int opacityActive = 255;
 
-    // update led mask
-    void updateLedMask(const int wordMask[], int arraySize)
-    {
-      // loop over array
-      for (int x=0; x < arraySize; x++) 
-      {
-        // check if mask has a valid LED number
-        if (wordMask[x] >= 0 && wordMask[x] < maskSizeLeds)
-        {
-          // turn LED on
-          maskLedsOn[wordMask[x]] = 1;
-        }
-      }
+  // Opacity (0=off, 255=full brightness) applied to LEDs that are NOT part of the current time sentence.
+  int opacityInactive = 0;
+
+  // Number of characters to skip at the start of the matrix when mapping to physical LEDs.
+  // E.g. set to 10 to skip the first row, so matrix char 10 maps to physical LED 0.
+  int matrixCharOffset = 0;
+
+  // Test time override: set testHour (0‥23) and testMinute (0‥59) to force a specific time to be
+  // displayed instead of the real time. Set testHour to -1 to disable (use real time).
+  int testHour = -1;
+  int testMinute = 0;
+
+  void allocateLedMask() {
+    if (ledMask) {
+      d_free(ledMask);
+      ledMask = nullptr;
     }
 
-    // set hours
-    void setHours(int hours, bool fullClock)
-    {
-      int index = hours;
+    size_t matrixLength = characterMatrix.length();
 
-      // handle 00:xx as 12:xx
-      if (hours == 0)
-      {
-        index = 12;
-      }
+    if (matrixLength == 0)
+      return;
 
-      // check if we get an overrun of 12 o´clock
-      if (hours == 13)
-      {
-        index = 1;
-      }
+    ledMask = (bool*) d_malloc(matrixLength * sizeof(bool));
 
-      // special handling for "ein Uhr" instead of "eins Uhr"
-      if (hours == 1 && fullClock == true)
-      {
-        index = 0;
-      }
+    if (ledMask)
+      memset(ledMask, 0, matrixLength * sizeof(bool));
+  }
 
-      // update led mask
-      if (meander)
-      {
-        updateLedMask(maskHoursMea[index], maskSizeHoursMea);
-      } else {
-      updateLedMask(maskHours[index], maskSizeHours);
-      }
+  String lastSentence = "";
+
+  /*
+   * Update the ledMask for the current time, by setting ledMask[i] to true if
+   * the LED at index i should be on for the current time, and false if it
+   * should be off.
+   */
+  // Clamp a value to the inclusive [lo, hi] range.
+  int clampInt(int v, int lo, int hi) {
+    return v < lo ? lo : (v > hi ? hi : v);
+  }
+
+  void updateLedMaskForCurrentTime() {
+    const int currentMinutes = testHour >= 0
+      ? (testHour * 60 + testMinute) % 1440
+      : (hour(localTime) * 60 + minute(localTime)) % 1440;
+    WordClockCore::MinuteDotMarkers markers;
+
+    if (!WordClockCore::parseMinuteDotMarkers(characterMatrix.c_str(), characterMatrix.length(), markers))
+      return;
+
+    const WordClockCore::TimeContext time = WordClockCore::makeTimeContext(currentMinutes, markers.enabled());
+    WordClockCore::DisplayPlan plan;
+
+    bool placementOk = false;
+  #if defined(WORD_CLOCK_LANGUAGE_NL)
+    placementOk = WordClock::buildPlan(time, plan) &&
+            WordClock::placePlan(time, plan, characterMatrix, characterMatrixWidth, meander, ledMask);
+  #else
+    placementOk = WordClock::buildPlan(time, displayItIs, nord, plan) &&
+            WordClock::placePlan(time, plan, meander, ledMask, characterMatrix.length());
+  #endif
+    if (!placementOk)
+      return;
+
+    String sentence;
+
+    for (uint8_t index = 0; index < plan.count; ++index) {
+      if (index > 0) sentence += ' ';
+      sentence += FPSTR(WordClock::wordText(static_cast<WordClock::WordId>(plan.units[index].id)));
     }
 
-    // set minutes
-    void setMinutes(int index)
-    {
-      // update led mask
-      if (meander)
-      {
-        updateLedMask(maskMinutesMea[index], maskSizeMinutesMea);
-      } else {
-      updateLedMask(maskMinutes[index], maskSizeMinutes);
-      }
+    lastSentence = sentence;
+  };
+
+public:
+  // Functions called by WLED
+
+  ~WordClockUsermod() {
+    if (ledMask)
+      d_free(ledMask);
+  }
+
+  /*
+   * setup() is called once at boot. WiFi is not yet connected at this point.
+   * You can use it to initialize variables, sensors or similar.
+   */
+  void setup() {
+    allocateLedMask();
+  }
+
+  /*
+   * connected() is called every time the WiFi is (re)connected
+   * Use it to initialize network interfaces
+   */
+  void connected() {
+  }
+
+  /*
+   * loop() is called continuously. Here you can check for events, read sensors, etc.
+   *
+   * Tips:
+   * 1. You can use "if (WLED_CONNECTED)" to check for a successful network connection.
+   *    Additionally, "if (WLED_MQTT_CONNECTED)" is available to check for a connection to an MQTT broker.
+   *
+   * 2. Try to avoid using the delay() function. NEVER use delays longer than 10 milliseconds.
+   *    Instead, use a timer check as shown here.
+   */
+  void loop() {
+    // Execute only once per minute
+    if (millis() - lastTime < 60 * 1000)
+      return;
+
+    updateLedMaskForCurrentTime();
+
+    // Remember this update
+    lastTime = millis();
+  }
+
+  /*
+   * addToJsonInfo() can be used to add custom entries to the /json/info part of the JSON API.
+   * Creating an "u" object allows you to add custom key/value pairs to the Info section of the WLED web UI.
+   * Below it is shown how this could be used for e.g. a light sensor
+   */
+  void addToJsonInfo(JsonObject& root) {
+    JsonObject user = root["u"];
+    if (user.isNull()) user = root.createNestedObject("u");
+
+    // Current localTime so you can verify NTP has synced
+    char timeBuf[16];
+    snprintf(timeBuf, sizeof(timeBuf), "%02d:%02d:%02d",
+             hour(localTime), minute(localTime), second(localTime));
+    user[F("WClock localTime")] = timeBuf;
+
+    // The Dutch sentence currently displayed. Skip the constant "HET IS " prefix (7 chars)
+    // so the two visible characters in the UI show the variable part (e.g. "TIEN OVER DRIE").
+    user[F("WClock sentence")] = lastSentence.length() > 7 ? lastSentence.substring(7) : (lastSentence.isEmpty() ? F("(not computed yet)") : lastSentence);
+
+    // How many matrix LEDs are currently lit
+    int litCount = 0;
+
+    if (ledMask) {
+      for (int i = 0; i < (int)characterMatrix.length(); i++)
+        if (ledMask[i]) litCount++;
     }
 
-    // set minutes dot
-    void setSingleMinuteDots(int minutes)
-    {
-      // modulo to get minute dots
-      int minutesDotCount = minutes % 5;
+    user[F("WClock lit LEDs")] = litCount;
+  }
 
-      // check if minute dots are active
-      if (minutesDotCount > 0)
-      {
-        // activate all minute dots until number is reached
-        for (int i = 0; i < minutesDotCount; i++)
-        {
-          // activate LED
-          maskLedsOn[maskMinuteDots[i]] = 1;  
-        }
-      }
+  /*
+   * addToJsonState() can be used to add custom entries to the /json/state part of the JSON API (state object).
+   * Values in the state object may be modified by connected clients
+   */
+  void addToJsonState(JsonObject &root) {
+  }
+
+  /*
+   * readFromJsonState() can be used to receive data clients send to the /json/state part of the JSON API (state object).
+   * Values in the state object may be modified by connected clients
+   */
+  void readFromJsonState(JsonObject &root) {
+  }
+
+  /*
+   * addToConfig() can be used to add custom persistent settings to the cfg.json file in the "um" (usermod) object.
+   * It will be called by WLED when settings are actually saved (for example, LED settings are saved)
+   * If you want to force saving the current state, use serializeConfig() in your loop().
+   *
+   * CAUTION: serializeConfig() will initiate a filesystem write operation.
+   * It might cause the LEDs to stutter and will cause flash wear if called too often.
+   * Use it sparingly and always in the loop, never in network callbacks!
+   *
+   * addToConfig() will make your settings editable through the Usermod Settings page automatically.
+   *
+   * Usermod Settings Overview:
+   * - Numeric values are treated as floats in the browser.
+   *   - If the numeric value entered into the browser contains a decimal point, it will be parsed as a C float
+   *     before being returned to the Usermod.  The float data type has only 6-7 decimal digits of precision, and
+   *     doubles are not supported, numbers will be rounded to the nearest float value when being parsed.
+   *     The range accepted by the input field is +/- 1.175494351e-38 to +/- 3.402823466e+38.
+   *   - If the numeric value entered into the browser doesn't contain a decimal point, it will be parsed as a
+   *     C int32_t (range: -2147483648 to 2147483647) before being returned to the usermod.
+   *     Overflows or underflows are truncated to the max/min value for an int32_t, and again truncated to the type
+   *     used in the Usermod when reading the value from ArduinoJson.
+   * - Pin values can be treated differently from an integer value by using the key name "pin"
+   *   - "pin" can contain a single or array of integer values
+   *   - On the Usermod Settings page there is simple checking for pin conflicts and warnings for special pins
+   *     - Red color indicates a conflict.  Yellow color indicates a pin with a warning (e.g. an input-only pin)
+   *   - Tip: use int8_t to store the pin value in the Usermod, so a -1 value (pin not set) can be used
+   *
+   * See usermod_v2_auto_save.h for an example that saves Flash space by reusing ArduinoJson key name strings
+   *
+   * If you need a dedicated settings page with custom layout for your Usermod, that takes a lot more work.
+   * You will have to add the setting to the HTML, xml.cpp and set.cpp manually.
+   * See the WLED Soundreactive fork (code and wiki) for reference.  https://github.com/atuline/WLED
+   *
+   * I highly recommend checking out the basics of ArduinoJson serialization and deserialization in order to use custom settings!
+   */
+  void addToConfig(JsonObject &root) {
+    JsonObject top = root.createNestedObject(F("WordClockUsermod"));
+    top[F("active")] = usermodActive;
+    top[F("displayItIs")] = displayItIs;
+    top[F("ledOffset")] = ledOffset;
+    top[F("Norddeutsch")] = nord;
+    top[F("Brightness_Active")] = opacityActive;
+    top[F("Brightness_Inactive")] = opacityInactive;
+    top[F("meander")] = meander;
+    top[F("Character_Matrix")] = characterMatrix;
+    top[F("Character_Matrix_Width")] = characterMatrixWidth;
+    top[F("Matrix_Char_Offset")] = matrixCharOffset;
+    top[F("Test_Hour")] = testHour;
+    top[F("Test_Minute")] = testMinute;
+  }
+
+  void appendConfigData() {
+    // Add hints for the Usermod Settings page, so the user knows what the settings mean
+    oappend(F("addInfo('WordClockUsermod:Brightness_Active', 1, '(0-255)');"));
+    oappend(F("addInfo('WordClockUsermod:Brightness_Inactive', 1, '(0-255)');"));
+    oappend(F("addInfo('WordClockUsermod:Test_Hour', 1, '(0-23, -1 for real time)');"));
+    oappend(F("addInfo('WordClockUsermod:Test_Minute', 1, '(0-59)');"));
+    oappend(F("addInfo('WordClockUsermod:Norddeutsch', 1, 'Viertel vor instead of Dreiviertel');"));
+  }
+
+  /*
+   * readFromConfig() can be used to read back the custom settings you added with addToConfig().
+   * This is called by WLED when settings are loaded (currently this only happens immediately after boot, or after saving on the Usermod Settings page)
+   *
+   * readFromConfig() is called BEFORE setup(). This means you can use your persistent values in setup() (e.g. pin assignments, buffer sizes),
+   * but also that if you want to write persistent values to a dynamic buffer, you'd need to allocate it here instead of in setup.
+   * If you don't know what that is, don't fret. It most likely doesn't affect your use case :)
+   *
+   * Return true in case the config values returned from Usermod Settings were complete, or false if you'd like WLED to save your defaults to disk (so any missing values are editable in Usermod Settings)
+   *
+   * getJsonValue() returns false if the value is missing, or copies the value into the variable provided and returns true if the value is present
+   * The configComplete variable is true only if the "exampleUsermod" object and all values are present.  If any values are missing, WLED will know to call addToConfig() to save them
+   *
+   * This function is guaranteed to be called on boot, but could also be called every time settings are updated
+   */
+  bool readFromConfig(JsonObject &root) {
+    // default settings values could be set here (or below using the 3-argument getJsonValue()) instead of in the class definition or constructor
+    // setting them inside readFromConfig() is slightly more robust, handling the rare but plausible use case of single value being missing after boot (e.g. if the cfg.json was manually edited and a value was removed)
+
+    JsonObject top = root[F("WordClockUsermod")];
+    bool legacyConfig = top.isNull();
+    if (legacyConfig)
+      top = root[F("Word Clock NL")];
+
+    bool configComplete = !top.isNull() && !legacyConfig;
+
+    configComplete &= getJsonValue(top[F("active")], usermodActive);
+    getJsonValue(top[F("displayItIs")], displayItIs);
+    getJsonValue(top[F("ledOffset")], ledOffset);
+    getJsonValue(top[F("Norddeutsch")], nord);
+    getJsonValue(top[F("Brightness_Active")], opacityActive);
+    getJsonValue(top[F("Brightness_Inactive")], opacityInactive);
+    opacityActive = clampInt(opacityActive, 0, 255);
+    opacityInactive = clampInt(opacityInactive, 0, 255);
+    bool prevMeander = meander;
+    getJsonValue(top[F("meander")], meander);
+
+    if (meander != prevMeander) {
+      lastSentence = "";             // force mask recompute
+      lastTime = ULONG_MAX - 60000UL; // trigger recompute on very next loop() call
     }
 
-    // update the display
-    void updateDisplay(uint8_t hours, uint8_t minutes) 
-    {
-      // disable complete matrix at the bigging
-      for (int x = 0; x < maskSizeLeds; x++)
-      {
-        maskLedsOn[x] = 0;
-      } 
-      
-      // display it is/es ist if activated
-      if (displayItIs)
-      {
-        updateLedMask(maskItIs, maskSizeItIs);
-      }
+    String prevCharacterMatrix = characterMatrix;
+    getJsonValue(top[F("Character_Matrix")], characterMatrix);
 
-      // set single minute dots
-      setSingleMinuteDots(minutes);
+    if (!characterMatrix.equals(prevCharacterMatrix)) {
+      allocateLedMask();
 
-      // switch minutes
-      switch (minutes / 5) 
-      {
-        case 0:
-            // full hour
-            setMinutes(0);
-            setHours(hours, true);
-            break;
-        case 1:
-            // 5 nach
-            setMinutes(1);
-            setHours(hours, false);
-            break;
-        case 2:
-            // 10 nach
-            setMinutes(2);
-            setHours(hours, false);
-            break;
-        case 3:
-            if (nord) {
-              // viertel nach
-              setMinutes(12);
-              setHours(hours, false);
-            } else {
-              // viertel 
-              setMinutes(3);
-              setHours(hours + 1, false);
-            };
-            break;
-        case 4:
-            // 20 nach
-            setMinutes(4);
-            setHours(hours, false);
-            break;
-        case 5:
-            // 5 vor halb
-            setMinutes(5);
-            setHours(hours + 1, false);
-            break;
-        case 6:
-            // halb
-            setMinutes(6);
-            setHours(hours + 1, false);
-            break;
-        case 7:
-            // 5 nach halb
-            setMinutes(7);
-            setHours(hours + 1, false);
-            break;
-        case 8:
-            // 20 vor
-            setMinutes(8);
-            setHours(hours + 1, false);
-            break;
-        case 9:
-            // viertel vor
-            if (nord) {
-              setMinutes(13);
-            } 
-            // dreiviertel
-              else {
-              setMinutes(9);
-            }
-            setHours(hours + 1, false);
-            break;
-        case 10:
-            // 10 vor
-            setMinutes(10);
-            setHours(hours + 1, false);
-            break;
-        case 11:
-            // 5 vor
-            setMinutes(11);
-            setHours(hours + 1, false);
-            break;
-        }
+      lastSentence = "";             // force mask recompute
+      lastTime = ULONG_MAX - 60000UL; // trigger recompute on very next loop() call
     }
 
-  public:
-    //Functions called by WLED
+    int prevCharacterMatrixWidth = characterMatrixWidth;
+    getJsonValue(top[F("Character_Matrix_Width")], characterMatrixWidth);
+    characterMatrixWidth = clampInt(characterMatrixWidth, WordClock::maxWordLength(), characterMatrix.length());
 
-    /*
-     * setup() is called once at boot. WiFi is not yet connected at this point.
-     * You can use it to initialize variables, sensors or similar.
-     */
-    void setup() 
-    {
+    if (characterMatrixWidth != prevCharacterMatrixWidth) {
+      lastSentence = "";             // force mask recompute
+      lastTime = ULONG_MAX - 60000UL; // trigger recompute on very next loop() call
     }
 
-    /*
-     * connected() is called every time the WiFi is (re)connected
-     * Use it to initialize network interfaces
-     */
-    void connected() 
-    {
+    getJsonValue(top[F("Matrix_Char_Offset")], matrixCharOffset);
+    matrixCharOffset = clampInt(matrixCharOffset, 0, characterMatrix.length());
+
+    int prevTestHour = testHour;
+    int prevTestMinute = testMinute;
+    getJsonValue(top[F("Test_Hour")], testHour);
+    getJsonValue(top[F("Test_Minute")], testMinute);
+    testHour = clampInt(testHour, -1, 23);
+    testMinute = clampInt(testMinute, 0, 59);
+
+    if (testHour != prevTestHour || testMinute != prevTestMinute) {
+      lastSentence = "";               // force mask recompute
+      lastTime = ULONG_MAX - 60000UL;  // trigger recompute on very next loop() call
     }
 
-    /*
-     * loop() is called continuously. Here you can check for events, read sensors, etc.
-     * 
-     * Tips:
-     * 1. You can use "if (WLED_CONNECTED)" to check for a successful network connection.
-     *    Additionally, "if (WLED_MQTT_CONNECTED)" is available to check for a connection to an MQTT broker.
-     * 
-     * 2. Try to avoid using the delay() function. NEVER use delays longer than 10 milliseconds.
-     *    Instead, use a timer check as shown here.
-     */
-    void loop() {
+    return configComplete;
+  }
 
-      // do it every 5 seconds
-      if (millis() - lastTime > 5000) 
-      {
-        // check the time
-        int minutes = minute(localTime);
+  /*
+   * handleOverlayDraw() is called just before every show() (LED strip update frame) after effects have set the colors.
+   * Use this to blank out some LEDs or set them to a different color regardless of the set effect mode.
+   * Commonly used for custom clocks (Cronixie, 7 segment)
+   */
+  void handleOverlayDraw() {
+    // Check if usermod is active
+    if (!usermodActive)
+      return;
 
-        // check if we already updated this minute
-        if (lastTimeMinutes != minutes)
-        {
-          // update the display with new time
-          updateDisplay(hourFormat12(localTime), minute(localTime));
+    if (!ledMask)
+      return;
 
-          // remember last update time
-          lastTimeMinutes = minutes;
-        }
+    int matrixLen = (int)characterMatrix.length();
 
-        // remember last update
-        lastTime = millis();
-      }
+    // Loop over all leds
+    for (int i = matrixCharOffset; i < matrixLen; i++) {
+      int physIndex = ledOffset + i - matrixCharOffset;
+      uint32_t color = strip.getPixelColor(physIndex);
+      // Scale by opacityActive for lit LEDs, opacityInactive for dimmed LEDs.
+      int scale = ledMask[i] ? opacityActive : opacityInactive;
+      uint8_t r = ((color >> 16) & 0xFF) * scale / 255;
+      uint8_t g = ((color >>  8) & 0xFF) * scale / 255;
+      uint8_t b = ((color >>  0) & 0xFF) * scale / 255;
+      strip.setPixelColor(physIndex, RGBW32(r, g, b, 0));
     }
+  }
 
-    /*
-     * addToJsonInfo() can be used to add custom entries to the /json/info part of the JSON API.
-     * Creating an "u" object allows you to add custom key/value pairs to the Info section of the WLED web UI.
-     * Below it is shown how this could be used for e.g. a light sensor
-     */
-    /*
-    void addToJsonInfo(JsonObject& root)
-    {
-    }
-    */
+  /*
+   * getId() allows you to optionally give your V2 usermod an unique ID (please define it in const.h!).
+   * This could be used in the future for the system to determine whether your usermod is installed.
+   */
+  uint16_t getId() {
+    return USERMOD_ID_WORDCLOCK;
+  }
 
-    /*
-     * addToJsonState() can be used to add custom entries to the /json/state part of the JSON API (state object).
-     * Values in the state object may be modified by connected clients
-     */
-    void addToJsonState(JsonObject& root)
-    {
-    }
-
-    /*
-     * readFromJsonState() can be used to receive data clients send to the /json/state part of the JSON API (state object).
-     * Values in the state object may be modified by connected clients
-     */
-    void readFromJsonState(JsonObject& root)
-    {
-    }
-
-    /*
-     * addToConfig() can be used to add custom persistent settings to the cfg.json file in the "um" (usermod) object.
-     * It will be called by WLED when settings are actually saved (for example, LED settings are saved)
-     * If you want to force saving the current state, use serializeConfig() in your loop().
-     * 
-     * CAUTION: serializeConfig() will initiate a filesystem write operation.
-     * It might cause the LEDs to stutter and will cause flash wear if called too often.
-     * Use it sparingly and always in the loop, never in network callbacks!
-     * 
-     * addToConfig() will make your settings editable through the Usermod Settings page automatically.
-     *
-     * Usermod Settings Overview:
-     * - Numeric values are treated as floats in the browser.
-     *   - If the numeric value entered into the browser contains a decimal point, it will be parsed as a C float
-     *     before being returned to the Usermod.  The float data type has only 6-7 decimal digits of precision, and
-     *     doubles are not supported, numbers will be rounded to the nearest float value when being parsed.
-     *     The range accepted by the input field is +/- 1.175494351e-38 to +/- 3.402823466e+38.
-     *   - If the numeric value entered into the browser doesn't contain a decimal point, it will be parsed as a
-     *     C int32_t (range: -2147483648 to 2147483647) before being returned to the usermod.
-     *     Overflows or underflows are truncated to the max/min value for an int32_t, and again truncated to the type
-     *     used in the Usermod when reading the value from ArduinoJson.
-     * - Pin values can be treated differently from an integer value by using the key name "pin"
-     *   - "pin" can contain a single or array of integer values
-     *   - On the Usermod Settings page there is simple checking for pin conflicts and warnings for special pins
-     *     - Red color indicates a conflict.  Yellow color indicates a pin with a warning (e.g. an input-only pin)
-     *   - Tip: use int8_t to store the pin value in the Usermod, so a -1 value (pin not set) can be used
-     *
-     * See usermod_v2_auto_save.h for an example that saves Flash space by reusing ArduinoJson key name strings
-     * 
-     * If you need a dedicated settings page with custom layout for your Usermod, that takes a lot more work.  
-     * You will have to add the setting to the HTML, xml.cpp and set.cpp manually.
-     * See the WLED Soundreactive fork (code and wiki) for reference.  https://github.com/atuline/WLED
-     * 
-     * I highly recommend checking out the basics of ArduinoJson serialization and deserialization in order to use custom settings!
-     */
-    void addToConfig(JsonObject& root)
-    {
-      JsonObject top = root.createNestedObject(F("WordClockUsermod"));
-      top[F("active")] = usermodActive;
-      top[F("displayItIs")] = displayItIs;
-      top[F("ledOffset")] = ledOffset;
-      top[F("Meander wiring?")] = meander;
-      top[F("Norddeutsch")] = nord;
-    }
-
-    void appendConfigData()
-    {
-      oappend(F("addInfo('WordClockUsermod:ledOffset', 1, 'Number of LEDs before the letters');"));
-      oappend(F("addInfo('WordClockUsermod:Norddeutsch', 1, 'Viertel vor instead of Dreiviertel');"));
-    }
-
-    /*
-     * readFromConfig() can be used to read back the custom settings you added with addToConfig().
-     * This is called by WLED when settings are loaded (currently this only happens immediately after boot, or after saving on the Usermod Settings page)
-     * 
-     * readFromConfig() is called BEFORE setup(). This means you can use your persistent values in setup() (e.g. pin assignments, buffer sizes),
-     * but also that if you want to write persistent values to a dynamic buffer, you'd need to allocate it here instead of in setup.
-     * If you don't know what that is, don't fret. It most likely doesn't affect your use case :)
-     * 
-     * Return true in case the config values returned from Usermod Settings were complete, or false if you'd like WLED to save your defaults to disk (so any missing values are editable in Usermod Settings)
-     * 
-     * getJsonValue() returns false if the value is missing, or copies the value into the variable provided and returns true if the value is present
-     * The configComplete variable is true only if the "exampleUsermod" object and all values are present.  If any values are missing, WLED will know to call addToConfig() to save them
-     * 
-     * This function is guaranteed to be called on boot, but could also be called every time settings are updated
-     */
-    bool readFromConfig(JsonObject& root)
-    {
-      // default settings values could be set here (or below using the 3-argument getJsonValue()) instead of in the class definition or constructor
-      // setting them inside readFromConfig() is slightly more robust, handling the rare but plausible use case of single value being missing after boot (e.g. if the cfg.json was manually edited and a value was removed)
-
-      JsonObject top = root[F("WordClockUsermod")];
-
-      bool configComplete = !top.isNull();
-
-      configComplete &= getJsonValue(top[F("active")], usermodActive);
-      configComplete &= getJsonValue(top[F("displayItIs")], displayItIs);
-      configComplete &= getJsonValue(top[F("ledOffset")], ledOffset);
-      configComplete &= getJsonValue(top[F("Meander wiring?")], meander);
-      configComplete &= getJsonValue(top[F("Norddeutsch")], nord);
-
-      return configComplete;
-    }
-
-    /*
-     * handleOverlayDraw() is called just before every show() (LED strip update frame) after effects have set the colors.
-     * Use this to blank out some LEDs or set them to a different color regardless of the set effect mode.
-     * Commonly used for custom clocks (Cronixie, 7 segment)
-     */
-    void handleOverlayDraw()
-    {
-      // check if usermod is active
-      if (usermodActive == true)
-      {
-        // loop over all leds
-        for (int x = 0; x < maskSizeLeds; x++)
-        {
-          // check mask
-          if (maskLedsOn[x] == 0)
-          {
-            // set pixel off
-            strip.setPixelColor(x + ledOffset, RGBW32(0,0,0,0));
-          }
-        }
-      }
-    }
-
-    /*
-     * getId() allows you to optionally give your V2 usermod an unique ID (please define it in const.h!).
-     * This could be used in the future for the system to determine whether your usermod is installed.
-     */
-    uint16_t getId()
-    {
-      return USERMOD_ID_WORDCLOCK;
-    }
-
-   //More methods can be added in the future, this example will then be extended.
-   //Your usermod will remain compatible as it does not need to implement all methods from the Usermod base class!
+  // More methods can be added in the future, this example will then be extended.
+  // Your usermod will remain compatible as it does not need to implement all methods from the Usermod base class!
 };
 
 static WordClockUsermod usermod_v2_word_clock;
