--- usermods/usermod_v2_word_clock_nl/library.json
@@ -1,4 +1,0 @@
-{
-  "name": "usermod_v2_word_clock_nl",
-  "build": { "libArchive": false }
-}
