+++ wled00/const.h
@@ -252,7 +252,6 @@ static_assert(WLED_MAX_BUSSES <= 32, "WLED_MAX_BUSSES exceeds hard limit");
 #define USERMOD_ID_RF433                 56     //Usermod "usermod_v2_RF433.h"
 #define USERMOD_ID_BRIGHTNESS_FOLLOW_SUN 57     //Usermod "usermod_v2_brightness_follow_sun.h"
 #define USERMOD_ID_USER_FX               58     //Usermod "user_fx"
-#define USERMOD_ID_WORDCLOCK_NL          59     //Usermod "usermod_v2_word_clock_nl"
 
 //Wifi encryption type
 #ifdef WLED_ENABLE_WPA_ENTERPRISE
